requirements.txt
streamlit==1.39.0
requests
python-dotenv
tqdm